[tests] Move Snowball support into Marin
