+++ lib/marin/src/marin/testing/__init__.py
@@ -1,2 +1,4 @@
 # Copyright The Marin Authors
 # SPDX-License-Identifier: Apache-2.0
+
+"""Reusable support for Marin tests and validation jobs."""
