+++ lib/marin/src/marin/testing/inference/__init__.py
@@ -1,2 +1,4 @@
 # Copyright The Marin Authors
 # SPDX-License-Identifier: Apache-2.0
+
+"""Reusable inference validation support."""
