+++ lib/marin/src/marin/testing/inference/resources/june_tpu_67b_a2b_step_42150_representative_eval_golden.json
