+++ tests/cluster/vllm/test_snowball_backend_parity.py
@@ -17,9 +17,8 @@ from fray.types import Entrypoint, JobRequest, ResourceConfig, create_environmen
 from iris.client import IrisClient
 from iris.cluster.setup_scripts import default_setup_script
 from iris.rpc import job_pb2
-
-from .snowball import RepresentativeGolden, read_representative_goldens
-from .snowball_backend_parity_jobs import (
+from marin.testing.inference.snowball import RepresentativeGolden, read_representative_goldens
+from marin.testing.inference.snowball_backend_parity_jobs import (
     GPU_COUNT,
     score_levanter_against_goldens,
     score_vllm_against_goldens,
